requirements.txt
streamlit>=1.20.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
requests>=2.25.0
openpyxl>=3.0.0
scipy>=1.7.0
scikit-learn>=1.0.0
python-binance>=1.0.17 